Correction for Defect 5949
At some point, E+ stopped producing some output variables for detached shading elements.  The problem was traced back to ExtSolar not being set to true for those surfaces.  Simply adding additional checks in the report set-up does not correctly address this problem; this was tested but not all of the solar output variables are calculated because they require ExtSolar to be set.