Fix for #10317 error
A subtle error in the logic in the swimming pool caused it to turn off when the boiler was not high enough capacity to get the water temperature above the pool setpoint.  The problem is if the pool has changing water setpoints and the boiler is not very high capacity, the outlet boiler temperature might not get the water to above the setpoint but could still provide heating because it is higher than the ACTUAL pool temperature.  In develop, this causes everything to shut down and the boiler never runs.  The fix gets the boiler to run and the pool slowly gets back to the setpoint temperature.