Fix for #5699
This makes it so that supporting input for the CondFD method via the objects MaterialProperty:PhaseChange and MaterialProperty:VariableThermalConductivity are extensible.  The main temporary array that is used to grab numeric input and then pass it to local data structures is now allocated based on the user input file rather than the old fixed limits or even the new extensible list (capped now at 100 based on the IDD changes).  Also includes an update schema file since two new items were made extensible.