Fix for Defect #10579
Implementation of a new feature to allow the user to allow the user to still use the Site:Location designation even when a weather file exists.  Currently, if a weather file exists, it always overwrite the Site:Location information.  This allows the user to choose.  Commit includes modifications to the IDD, code to read in the new input and handle it properly within the code, and a new input file that shows that the Site:Location information is kept when requested by the user.  Expecting no changes in existing IDFs since the default is to maintain the status quo from previous editions.