10065 Evaporator Temp Checks
Created a new routine that is used by both the chiller and chillerheater in this model.  This is again to improve clarity in the routines and prepare for the changes that will be needed to solve the actual defect which is that the evaporator and compressor energy never get updated when heating is the predominant mode (modes 2 and 5).