10299 Split GetShadowingInput
This commit splits GetShadowingInput into two separate routines and will test to see if the split has been done correctly.  The split is important because the actual getting of the input needs to be moved up in the order of calls to resolve this defect, but some of the processing needs to stay where it is at because it needs zones and surfaces read in.