10299 Separation of Get and Process
The previous version of GetShadowingInput obtained the ShadowCalculation input but then also did some processing.  In order to solve #10299, these needed to be separated so that when surfaces are read in that it knows the ShadowCalculation input and thus can produce error messages appropriately.  This commit separates the two parts of the previous routine into two routines which are now called at the appropriate places.  The next step will be to implement the solution and create a unit test.