10065 Fixes for ChillerHeater
This commit includes fixes for two separate issues.  First, when in heating mode, the evaporator and compressor power was always at full power, not at the power that the condenser needed.  Second, there were random times when in simultaneously heating and cooling mode that no evaporator load or compressor energy due to a logic flaw in the code.  This should fix those issues though there are probably more issues with this model.  This at least addresses those two issues that have caused users problems.