10065 Backed Out One Incorrent Mod
After further review of the code and the concepts of minPLR and frac, I realized that this one change was not consistent with the concept and the change was a mistake on my part.  This backs out this change and corrects the unit test as well.  Now this is PR ready.