change ElecPowerMinFlowRateFrac calc, autosize DesignElecPower
to compute "Electric Power Minimum Flow Rate Fraction", use the following logic

This field in Fan:SystemModel is "Electric Power Minimum Flow Rate Fraction", so
if the old fan is FixedFlowRate then
the new field should be old fanPowerMinAirFlow_str / maxAirFlow_str
If maxAirFlow_str is autosize, and fanPowerMinAirFlow_str` is > 0,
then throw a warning and set this field to 0

if the old fan is Fraction, then this field = old fan minAirFlowFrac_str.

For "Design Electric Power Consumption", directly make it autosize and use
TotalEfficiencyAndPressure in the sizing method.
This way also ensures the "_localTotalEff" variable in the code to use inputs
from "Fan Total Efficiency" in Fan:SystemModel