10638 Reorganization and Input File Mods
Reorganized some of the sizing routine into a new subroutine that covers both heating and cooling.  Also made two minor typo corrections in existing input files and made mods to one so that the code in question actually gets used (added design sizing specifications).  Only diffs would likely be in the file that has design specifications added since that was the source of the defect.