Patch for Q_cond_req < min speed power, CompSpdActual int to real
When Q_cond_req < ComoEvaporatingPWRSpd(1), Q_evap_req will be negative, the
calculated compressor speed will be negative. The case where Q_cond_req <
ComoEvaporatingPWRSpd(1) need to be considered, where compressor speed is
dependent on the compressour power curve output value.

CompSpdActual argument VRFOU_CompCap is an int, but the value passed in was
originally a real. This casting from int to real will truncate the value. If the
value is less than 1, it will be set to 0.