Cycling ratio need to be applied inside VRFOU_CalcCompH
because the demand is roughly Q_evap_req = TU_load + Pipe_Q - Ncomp (multiplied
by an adjustment factor C_cap_operation). Ncomp is an input-ouptut variable of
the function. Previously Cycling ratio is multiplied outside of VRFOU_CalcCompH
after the function has computed Ncomp. However, this will lead to very small
cycling ratio non-compatible with the demand as the demand is underestimated
because of the Ncomp (in reality it should not be this large, it should be
Ncomp * CyclingRatio). This doesn't matter on the cooling side as Ncomp is not
involved in the demand calculation there.